Put MapSlots tests in its own file
